+++ app/bootstrap.py
@@ -19,6 +19,7 @@ from app.modules.dashboard.controller import ROUTES as DASHBOARD_ROUTES
 from app.modules.dashboard.controller import dashboard_controller
 from app.modules.patients.controller import ROUTES as PATIENT_ROUTES
 from app.modules.registration.controller import ROUTES as REGISTRATION_ROUTES
+from app.modules.roles.controller import ROUTES as ROLES_ROUTES
 from app.modules.roles.service import user_has_permission
 from app.modules.settings.controller import ROUTES as SETTINGS_ROUTES
 from app.modules.visits.controller import ROUTES as VISIT_ROUTES
@@ -35,6 +36,7 @@ ROUTES = (
     + VISIT_ROUTES
     + CONSULTATION_ROUTES
     + SETTINGS_ROUTES
+    + ROLES_ROUTES
 )